requirements.txt
Django==3.2
django-cors-headers==3.5
djangorestframework==3.12
Pillow==8.2
django-environ==0.4.5
django-jazzmin==2.4.7
